chore(coin): remove transport dependencies from coin-framework
