+++ .changeset/orange-birds-wonder.md
@@ -1,0 +1,21 @@
+---
+"@ledgerhq/coin-algorand": minor
+"@ledgerhq/coin-polkadot": minor
+"@ledgerhq/coin-bitcoin": minor
+"@ledgerhq/coin-cardano": minor
+"@ledgerhq/coin-stellar": minor
+"@ledgerhq/coin-cosmos": minor
+"@ledgerhq/coin-elrond": minor
+"@ledgerhq/coin-solana": minor
+"@ledgerhq/coin-tezos": minor
+"@ledgerhq/coin-icon": minor
+"@ledgerhq/coin-near": minor
+"@ledgerhq/coin-tron": minor
+"@ledgerhq/coin-evm": minor
+"@ledgerhq/coin-ton": minor
+"@ledgerhq/coin-xrp": minor
+"@ledgerhq/live-common": minor
+"@ledgerhq/coin-framework": minor
+---
+
+Remove transport dependency from coin-framework
