+++ libs/coin-modules/coin-bitcoin/src/speculos-deviceActions.ts
@@ -1,10 +1,6 @@
 import type { DeviceAction } from "@ledgerhq/coin-framework/bot/types";
 import type { Transaction } from "./types";
-import {
-  deviceActionFlow,
-  formatDeviceAmount,
-  SpeculosButton,
-} from "@ledgerhq/coin-framework/bot/specs";
+import { deviceActionFlow, formatDeviceAmount, SpeculosButton } from "@ledgerhq/coin-framework/bot";
 import { perCoinLogic } from "./logic";
 
 export const acceptTransaction: DeviceAction<Transaction, any> = deviceActionFlow({
