+++ libs/coin-modules/coin-icon/src/speculos-deviceActions.ts
@@ -1,7 +1,7 @@
 import type { DeviceAction } from "@ledgerhq/coin-framework/bot/types";
 import type { Transaction } from "./types";
 import { formatCurrencyUnit } from "@ledgerhq/coin-framework/currencies/index";
-import { deviceActionFlow, SpeculosButton } from "@ledgerhq/coin-framework/bot/specs";
+import { deviceActionFlow, SpeculosButton } from "@ledgerhq/coin-framework/bot";
 import { getAccountCurrency } from "@ledgerhq/coin-framework/account/index";
 
 const confirmWording: Record<string, string> = {
