+++ libs/ledger-live-common/src/families/celo/speculos-deviceActions.ts
@@ -1,6 +1,7 @@
 import type { DeviceAction } from "../../bot/types";
 import type { Transaction } from "./types";
-import { deviceActionFlow, formatDeviceAmount, SpeculosButton } from "../../bot/specs";
+import { deviceActionFlow, formatDeviceAmount } from "../../bot/specs";
+import { SpeculosButton } from "../../bot/types";
 
 const typeWording = {
   send: "Send",
